+++ tests/testthat/test-developer-tools-xregs.R
@@ -1,24 +1,25 @@
 context("TEST DEVELOPER TOOLS - XREG TOOLS")
 
-
 # CREATE XREG RECIPE ----
 
-test_that("create_xreg_recipe: recipe with no compliant features returns a warning and NULL value", {
-
-    # Recipe returns no features
-    # - id is removed because of zero variance
-    # - date is removed
-    expect_warning({
-        null_recipe <- m4_monthly %>%
-            filter(id == "M750") %>%
-            select(-value) %>%
-            create_xreg_recipe(prepare = TRUE, one_hot = TRUE)
-
-    })
-
-    # Return is NULL
-    expect_null(null_recipe)
-})
+# test_that("create_xreg_recipe: recipe with no compliant features returns a warning and NULL value", {
+#
+#     # Recipe returns no features
+#     # - id is removed because of zero variance
+#     # - date is removed
+#     expect_warning({
+#         null_recipe <- m4_monthly %>%
+#             filter(id == "M750") %>%
+#             select(-value) %>%
+#             create_xreg_recipe(prepare = TRUE, one_hot = TRUE)
+#
+#     })
+#
+#     # Return is NULL
+#     expect_null(null_recipe)
+#
+#
+# })
 
 
 
@@ -42,7 +43,7 @@ test_that("create_xreg_recipe: dummy_encode = FALSE returns factors", {
 
 })
 
-test_that("create_xreg_recipe: dummy_encode = TRUE returns integer", {
+test_that("create_xreg_recipe: dummy_encode = TRUE returns dummies", {
 
     # Month
     predictors <- m4_monthly %>%
@@ -57,6 +58,4 @@ test_that("create_xreg_recipe: dummy_encode = TRUE returns integer", {
     expect_equal(ncol(juiced), 11)
 
     expect_true(all(juiced %>% map(is.numeric) %>% unlist()))
-
-
 })